Rag_1/requirements.txt
chromadb==1.0.20
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-openai==0.3.32
langchain-text-splitters==0.3.11
langdetect==1.0.9
langsmith==0.4.25
onnxruntime==1.22.1
openai==1.106.1
tiktoken==0.11.0
unstructured==0.18.14
unstructured-client==0.42.3
python-dotenv==1.0.1 
